LPS-115709 Rename classes
