+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/java/com/liferay/app/builder/workflow/web/internal/portlet/action/AddWorkflowAppMVCResourceCommand.java
@@ -15,7 +15,9 @@
 package com.liferay.app.builder.workflow.web.internal.portlet.action;
 
 import com.liferay.app.builder.constants.AppBuilderPortletKeys;
+import com.liferay.app.builder.rest.dto.v1_0.App;
 import com.liferay.app.builder.rest.resource.v1_0.AppResource;
+import com.liferay.app.builder.workflow.rest.dto.v1_0.AppWorkflow;
 import com.liferay.app.builder.workflow.rest.resource.v1_0.AppWorkflowResource;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
@@ -34,15 +36,15 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	property = {
 		"javax.portlet.name=" + AppBuilderPortletKeys.APPS,
-		"mvc.command.name=/app_builder/delete_workflow_app"
+		"mvc.command.name=/app_builder/add_workflow_app"
 	},
 	service = MVCResourceCommand.class
 )
-public class DeleteAppBuilderAppMVCResourceCommand
-	extends BaseAppBuilderAppMVCResourceCommand<Void> {
+public class AddWorkflowAppMVCResourceCommand
+	extends BaseAppBuilderAppMVCResourceCommand<App> {
 
 	@Override
-	protected Optional<Void> doTransactionalCommand(
+	protected Optional<App> doTransactionalCommand(
 			ResourceRequest resourceRequest)
 		throws Exception {
 
@@ -54,18 +56,21 @@ public class DeleteAppBuilderAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		appResource.deleteApp(
-			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
+		App app = appResource.postDataDefinitionApp(
+			ParamUtil.getLong(resourceRequest, "dataDefinitionId"),
+			App.toDTO(ParamUtil.getString(resourceRequest, "app")));
 
 		AppWorkflowResource appWorkflowResource = AppWorkflowResource.builder(
 		).user(
 			themeDisplay.getUser()
 		).build();
 
-		appWorkflowResource.deleteAppWorkflow(
-			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
+		appWorkflowResource.postAppWorkflow(
+			app.getId(),
+			AppWorkflow.toDTO(
+				ParamUtil.getString(resourceRequest, "appWorkflow")));
 
-		return Optional.empty();
+		return Optional.of(app);
 	}
 
 }
