+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/java/com/liferay/app/builder/workflow/web/internal/portlet/action/UpdateWorkflowAppMVCResourceCommand.java
@@ -15,8 +15,10 @@
 package com.liferay.app.builder.workflow.web.internal.portlet.action;
 
 import com.liferay.app.builder.constants.AppBuilderPortletKeys;
+import com.liferay.app.builder.model.AppBuilderAppVersion;
 import com.liferay.app.builder.rest.dto.v1_0.App;
 import com.liferay.app.builder.rest.resource.v1_0.AppResource;
+import com.liferay.app.builder.service.AppBuilderAppVersionLocalService;
 import com.liferay.app.builder.workflow.rest.dto.v1_0.AppWorkflow;
 import com.liferay.app.builder.workflow.rest.resource.v1_0.AppWorkflowResource;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
@@ -24,11 +26,13 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 
+import java.util.Objects;
 import java.util.Optional;
 
 import javax.portlet.ResourceRequest;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Rafael Praxedes
@@ -36,11 +40,11 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	property = {
 		"javax.portlet.name=" + AppBuilderPortletKeys.APPS,
-		"mvc.command.name=/app_builder/add_workflow_app"
+		"mvc.command.name=/app_builder/update_workflow_app"
 	},
 	service = MVCResourceCommand.class
 )
-public class AddAppBuilderAppMVCResourceCommand
+public class UpdateWorkflowAppMVCResourceCommand
 	extends BaseAppBuilderAppMVCResourceCommand<App> {
 
 	@Override
@@ -56,8 +60,13 @@ public class AddAppBuilderAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		App app = appResource.postDataDefinitionApp(
-			ParamUtil.getLong(resourceRequest, "dataDefinitionId"),
+		App app = appResource.getApp(
+			ParamUtil.getLong(resourceRequest, "appBuilderAppId"));
+
+		String appVersion = app.getVersion();
+
+		app = appResource.putApp(
+			ParamUtil.getLong(resourceRequest, "appBuilderAppId"),
 			App.toDTO(ParamUtil.getString(resourceRequest, "app")));
 
 		AppWorkflowResource appWorkflowResource = AppWorkflowResource.builder(
@@ -65,12 +74,52 @@ public class AddAppBuilderAppMVCResourceCommand
 			themeDisplay.getUser()
 		).build();
 
-		appWorkflowResource.postAppWorkflow(
-			app.getId(),
-			AppWorkflow.toDTO(
-				ParamUtil.getString(resourceRequest, "appWorkflow")));
+		AppWorkflow appWorkflow = AppWorkflow.toDTO(
+			ParamUtil.getString(resourceRequest, "appWorkflow"));
+
+		if (_isModifiedAppWorkflow(
+				appWorkflowResource.getAppWorkflow(
+					ParamUtil.getLong(resourceRequest, "appBuilderAppId")),
+				appWorkflow)) {
+
+			if (Objects.equals(app.getVersion(), appVersion)) {
+				AppBuilderAppVersion appBuilderAppVersion =
+					_appBuilderAppVersionLocalService.addAppBuilderAppVersion(
+						app.getSiteId(), themeDisplay.getCompanyId(),
+						themeDisplay.getUserId(), app.getId(),
+						app.getDataRecordCollectionId(),
+						app.getDataDefinitionId(), app.getDataLayoutId());
+
+				app.setVersion(appBuilderAppVersion.getVersion());
+			}
+
+			appWorkflowResource.putAppWorkflow(app.getId(), appWorkflow);
+		}
 
 		return Optional.of(app);
 	}
+
+	private boolean _isModifiedAppWorkflow(
+		AppWorkflow appWorkflow, AppWorkflow newAppWorkflow) {
+
+		if (!Objects.deepEquals(
+				appWorkflow.getAppWorkflowStates(),
+				newAppWorkflow.getAppWorkflowStates())) {
+
+			return true;
+		}
+
+		if (!Objects.deepEquals(
+				appWorkflow.getAppWorkflowTasks(),
+				newAppWorkflow.getAppWorkflowTasks())) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	@Reference
+	private AppBuilderAppVersionLocalService _appBuilderAppVersionLocalService;
 
 }
